refactor: social-detail:socialId 컴포넌트 파일 _components 폴더로 이동
