+++ src/app/auths/signup/page.tsx
@@ -4,7 +4,7 @@ import Link from 'next/link';
 import React from 'react';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import { SignUpFormData } from '@/types/user';
-import useCreateUser from '@/hooks/api/supabase/users/useCreateUser';
+import useCreateUser from '@/hooks/api/auth/useCreateUser';
 import { useRouter } from 'next/navigation';
 
 import InputForm from '@/components/common/Form/InputForm';
