+++ src/app/social/detail/[socialId]/_components/NotFoundRedirect.tsx
